+++ Word.py
@@ -11,5 +11,5 @@ class Word(Base):
     time_created = Column(DateTime(timezone=True), server_default=func.now())
     time_updated = Column(DateTime(timezone=True), onupdate=func.now())
     wordle_num = Column(Integer)
-    # date_used
+    date_used = Column(DateTime)